# api/swagger.yaml
swagger: "2.0"
info:
  description: "This is a simple bookstore server"
  version: "1.0.0"
  title: "BookStore"
  contact:
    email: "10530190@polimi.it"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5050"
basePath: "/v2"
...
schemes:
- "http"
paths:
...
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "userLoginPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data send to client"
        schema:
          type: "object"
          properties:
            email: 
              type: "string"
            name: 
              type: "string"
            password: 
              type: "string"
      responses:
        200:
          description: "Successful"
        404:
          description: "Unauthorized"
      x-swagger-router-controller: "User"
  /user/register:
    post:
      tags:
      - "user"
      summary: "Register"
      description: "Register into the store"
      operationId: "userRegisterPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data send to client"
        schema:
          type: "object"
          properties:
            email: 
              type: "string"
            name: 
              type: "string"
            password: 
              type: "string"
      responses:
        200:
          description: "Successful"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/{userId}:
    get:
      tags:
      - "user"
      summary: "Find user by ID"
      description: "Returns a user"
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/User'
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "User"      
...
  /carts:
    get:
      tags:
      - "cart"
      summary: "Find cart by ID"
      description: "Returns a list of books"
      operationId: "getCartById"
      produces:
      - "application/json"
      parameters:
      - name: "cartId"
        in: "query"
        description: "ID of book to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "A collection of Books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Cart"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cart not found"
      x-swagger-router-controller: "Cart"
  /cartInsert:
    post:
      tags:
      - "cart"
      summary: "cartInsert"
      description: "Register into the cart"
      operationId: "cartInsertPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data send to client"
        schema:
          type: "object"
          properties:
            user_id:
              type: "string"
            total:
              type: "string"
            quantity:
              type: "string"
            isbn:
              type: "string"             
      responses:
        200:
          description: "Successful"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "Cart"
  /cartDelete/{userId}:
    delete:
      tags:
      - "cart"
      description: "Returns books"
      operationId: "deleteCart"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Author"
        400:
          description: "Invalid ID supplied"
      x-swagger-router-controller: "Cart"
  /bookDelete/{userId}/{isbn}:
    delete:
      tags:
      - "cart"
      description: "Delete book"
      operationId: "deleteBook"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user"
        required: true
        type: "integer"
        format: "int64"
      - name: "isbn"
        in: "path"
        description: "isbn of book"
        required: true
        type: "string"        
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Author"
        400:
          description: "Invalid ID supplied"
      x-swagger-router-controller: "Cart"
definitions:
  Author:
    type: "object"
    required:
    - "id_author"
    - "name"
    - "biography"
    properties:
      id_author:
        type: "integer"
        format: "int64"
        $ref: "#/definitions/Relation"
      name:
        type: "string"
      biography:
        type: "string"
    title: "Author"
    description: "Author present in our website"
...
  User:
    type: "object"
    required:
      - "id_user"
      - "name"
      - "email"
    properties:
      id_user:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
    title: "User"
    description: "User in our database"
  Cart:
    type: "object"
    required:
      - "user_id"
      - "total"
      - "quantity"
      - "isbn"
    properties:
      user_id:
        type: "string"
      total:
        type: "string"
      quantity:
        type: "string"
      isbn:
        type: "string"
    title: "Cart"
    description: "Order for books"
  Book:
    type: "object"
    required:
    - "isbn"
    - "title"
    - "theme1"
    - "theme2"
    - "genre1"
    - "genre2"
    - "date"
    - "recommended"
    - "price"
    - "status"
    properties:
      isbn:
        type: "string"
        $ref: '#/definitions/Relation'
      title:
        type: "string"
      theme1:
        type: "string"
      theme2:
        type: "string"
      genre1:
        type: "string"
      genre2:
        type: "string"
      date:
        type: "string"
      recommended:
        type: "boolean"
      price:
        type: "number"  
      status:
        type: "string"
        description: "book availability in the inventory"
        enum:
        - "available"
        - "out of stock"
    title: "Book"
  Relation:
    type: "object"
    required:
      - "isbn"
      - "id_author"
    properties:
      id_author:
        type: "integer"
        format: "int64"
        $ref: "#/definitions/Author"
      isbn:
        type: "string"
        $ref: '#/definitions/Book'    
    description: "Relation between authors and books"